+++ drivers/rust/driver/src/lib.rs
@@ -2,10 +2,7 @@
 
 pub mod plugin_models;
 pub mod plugin_manager;
-#[cfg(not(windows))]
 mod child_process;
-#[cfg(windows)]
-mod child_process_windows;
 pub mod proto;
 pub mod catalogue_manager;
 pub mod content;
